requirements.txt
absl-py==2.1.0
altair==5.3.0
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deepface==0.0.92
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
gast==0.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.65.2
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.4.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
mtcnn==0.1.1
namex==0.0.8
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.4
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
retina-face==0.0.17
rich==13.7.1
rpds-py==0.19.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.17.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics>=8.0.122
urllib3==2.2.2
watchdog==4.0.1
Werkzeug==3.0.3
wrapt==1.16.0
